Crate grabbag_macros

Macros

collect!
This macro provides a way to initialise any container for which there is a FromIterator implementation. It allows for both sequence and map syntax to be used, as well as inline type ascription for the result.

count_exprs!
Counts the number of comma-delimited expressions passed to it. The result is a compile-time evaluable expression, suitable for use as a static array size, or the value of a const.

recurrence!
Expands to an expression implementing the Iterator trait, which yields successive elements of the given recurrence relationship.

sequence!
Expands to an expression implementing the Iterator trait, which yields successive elements of the given closed-form sequence.